Refactor code and add functions to be compatible with dot
Renew the limitations of the calculator because
numbers that are bigger than Number.MAX_SAFE_INTEGER
are not safe for calculating. There is a Big number as a solution
but the problem is that js displays numbers that are bigger than
1e20 are being displayed in exponential form which would have
been harder to work with.

Functions transformIntoCorrectForm, makeNewNumber, deleteLastDigit
were refactored to make it work with dot.

And some small error fixing