Fix previous commit, add clearing input
Made function clearNumber that accepts one parameter which is set to false
by default so that the function would only clear current entry, when it's set
to true - the history clears too.

In the previous commit the history used to be a span tag, but now
it's a pre tag so that when the text overflows it wouldn't
break lines, but it would grow horizontally.